Isolate diagnostic SSH probe flow
