+++ apps/mobile/src/lib/connection-diagnostic-runner.ts
@@ -2,6 +2,7 @@ import {
 	attemptSavedEntryWithTailscaleRecovery,
 	type SavedEntryTailscaleRecovery,
 } from './auto-connect-saved-entry';
+import { type ConnectAndOpenShellResult } from './connect-and-open-shell';
 import {
 	formatConnectionDiagnosticPrompt,
 	serializeConnectionDiagnosticError,
@@ -12,8 +13,6 @@ import {
 	type ConnectionDiagnosticTraceHandle,
 } from './connection-diagnostics';
 import { type SavedConnectionEntry } from './connection-utils';
-// eslint-disable-next-line import/consistent-type-specifier-style -- keep query-fns type-only so Node integration tests do not load React Native at runtime
-import type { ConnectAndOpenShellResult } from './query-fns';
 // eslint-disable-next-line import/consistent-type-specifier-style -- keep secrets-manager type-only so Node integration tests do not load React Native at runtime
 import type { InputConnectionDetails } from './secrets-manager';
 
@@ -42,10 +41,45 @@ export type ManualConnectionDiagnosticArgs = {
 	}) => Promise<ConnectAndOpenShellResult>;
 	recovery: SavedEntryTailscaleRecovery;
 	formatPrompt?: typeof formatConnectionDiagnosticPrompt;
+	timeoutMs?: number;
 };
+
+const DEFAULT_MANUAL_DIAGNOSTIC_TIMEOUT_MS = 60_000;
 
 let running = false;
 let activeTraceHandle: ConnectionDiagnosticTraceHandle | null = null;
+let activeRunToken: symbol | null = null;
+
+class ManualDiagnosticTimeoutError extends Error {
+	constructor(readonly timeoutMs: number) {
+		super(`Connection diagnostic timed out after ${timeoutMs}ms`);
+		this.name = 'ManualDiagnosticTimeoutError';
+	}
+}
+
+async function withManualDiagnosticTimeout<T>(
+	promise: Promise<T>,
+	timeoutMs: number,
+): Promise<T> {
+	let timeoutId: ReturnType<typeof setTimeout> | null = null;
+	try {
+		return await Promise.race([
+			promise,
+			new Promise<never>((_, reject) => {
+				timeoutId = setTimeout(() => {
+					timeoutId = null;
+					reject(new ManualDiagnosticTimeoutError(timeoutMs));
+				}, timeoutMs);
+				const maybeNodeTimer = timeoutId as ReturnType<typeof setTimeout> & {
+					unref?: () => void;
+				};
+				maybeNodeTimer.unref?.();
+			}),
+		]);
+	} finally {
+		if (timeoutId !== null) clearTimeout(timeoutId);
+	}
+}
 
 function getConnectionIdentity(
 	id: string,
@@ -116,105 +150,140 @@ export async function runManualConnectionDiagnostic(
 	}
 
 	running = true;
+	const runToken = Symbol('manual-connection-diagnostic');
+	activeRunToken = runToken;
 	let handle: ConnectionDiagnosticTraceHandle | null = null;
+	const ensureCurrentRun = () => {
+		if (activeRunToken !== runToken) {
+			throw new Error('Connection diagnostic run is no longer active');
+		}
+	};
 
 	try {
-		handle = args.recorder.startTrace({
-			trigger: 'manual-diagnostic',
-			reason: 'command-menu',
-		});
-		activeTraceHandle = handle;
-		const traceHandle = handle;
-		const latestEntry = await args.loadLatestSavedConnection();
-		if (!latestEntry) {
-			safeTraceEvent(traceHandle, {
-				type: 'manual-diagnostic.saved-entry.missing',
-				source: 'manual-diagnostic',
-				message: 'No eligible saved auto-connect connection was found.',
-			});
-			return finish(traceHandle, 'skipped', args);
-		}
+		return await withManualDiagnosticTimeout(
+			(async () => {
+				handle = args.recorder.startTrace({
+					trigger: 'manual-diagnostic',
+					reason: 'command-menu',
+				});
+				activeTraceHandle = handle;
+				const traceHandle = handle;
+				const latestEntry = await args.loadLatestSavedConnection();
+				ensureCurrentRun();
+				if (!latestEntry) {
+					safeTraceEvent(traceHandle, {
+						type: 'manual-diagnostic.saved-entry.missing',
+						source: 'manual-diagnostic',
+						message: 'No eligible saved auto-connect connection was found.',
+					});
+					return finish(traceHandle, 'skipped', args);
+				}
 
-		const details = latestEntry.value;
-		const normalizedDetails: InputConnectionDetails = {
-			...details,
-			useTmux: details.useTmux ?? true,
-			tmuxSessionName: details.tmuxSessionName?.trim() || 'main',
-			autoConnect: details.autoConnect ?? false,
-		};
-		const connection = getConnectionIdentity(latestEntry.id, normalizedDetails);
-		safeTraceEvent(traceHandle, {
-			type: 'manual-diagnostic.saved-entry.selected',
-			source: 'manual-diagnostic',
-			connection,
-		});
+				const details = latestEntry.value;
+				const normalizedDetails: InputConnectionDetails = {
+					...details,
+					useTmux: details.useTmux ?? true,
+					tmuxSessionName: details.tmuxSessionName?.trim() || 'main',
+					autoConnect: details.autoConnect ?? false,
+				};
+				const connection = getConnectionIdentity(
+					latestEntry.id,
+					normalizedDetails,
+				);
+				safeTraceEvent(traceHandle, {
+					type: 'manual-diagnostic.saved-entry.selected',
+					source: 'manual-diagnostic',
+					connection,
+				});
 
-		const resolvedSecurity = await args.resolveKeySecurity(details);
-		if (!resolvedSecurity) {
-			safeTraceEvent(traceHandle, {
-				type: 'manual-diagnostic.key-missing',
-				source: 'manual-diagnostic',
-				connection,
-			});
-			return finish(traceHandle, 'failed', args);
-		}
+				const resolvedSecurity = await args.resolveKeySecurity(details);
+				ensureCurrentRun();
+				if (!resolvedSecurity) {
+					safeTraceEvent(traceHandle, {
+						type: 'manual-diagnostic.key-missing',
+						source: 'manual-diagnostic',
+						connection,
+					});
+					return finish(traceHandle, 'failed', args);
+				}
 
-		safeTraceEvent(traceHandle, {
-			type: 'manual-diagnostic.key-resolved',
-			source: 'manual-diagnostic',
-			connection,
-		});
+				safeTraceEvent(traceHandle, {
+					type: 'manual-diagnostic.key-resolved',
+					source: 'manual-diagnostic',
+					connection,
+				});
 
-		const result = await attemptSavedEntryWithTailscaleRecovery({
-			platformOS: args.appState.platformOS,
-			recovery: args.recovery,
-			connectSavedEntry: () =>
-				args.connectSavedEntry({
-					connectionDetails: normalizedDetails,
-					resolvedSecurity,
+				const result = await attemptSavedEntryWithTailscaleRecovery({
+					platformOS: args.appState.platformOS,
+					recovery: args.recovery,
+					connectSavedEntry: () =>
+						Promise.resolve()
+							.then(ensureCurrentRun)
+							.then(() =>
+								args.connectSavedEntry({
+									connectionDetails: normalizedDetails,
+									resolvedSecurity,
+									trace: traceHandle,
+								}),
+							),
+					markTailscaleAttention: (message) => {
+						safeTraceEvent(traceHandle, {
+							type: 'manual-diagnostic.tailscale.attention',
+							source: 'tailscale-recovery',
+							message,
+						});
+					},
+					clearTailscaleAttention: () => {
+						safeTraceEvent(traceHandle, {
+							type: 'manual-diagnostic.tailscale.attention-cleared',
+							source: 'tailscale-recovery',
+						});
+					},
+					logTmuxAttachFailure: (tmuxResult) => {
+						safeTraceEvent(traceHandle, {
+							type: 'manual-diagnostic.tmux-attach-failed',
+							source: 'manual-diagnostic',
+							connection: {
+								connectionId: tmuxResult.connectionId,
+								tmuxSessionName: tmuxResult.tmuxSessionName,
+							},
+							details: {
+								tmuxAttachFailureReason: tmuxResult.tmuxAttachFailureReason,
+							},
+						});
+					},
+					logWarning: (message, error) => {
+						safeTraceEvent(traceHandle, {
+							type: 'manual-diagnostic.warning',
+							source: 'manual-diagnostic',
+							message,
+							error: serializeConnectionDiagnosticError(error),
+						});
+					},
 					trace: traceHandle,
-				}),
-			markTailscaleAttention: (message) => {
-				safeTraceEvent(traceHandle, {
-					type: 'manual-diagnostic.tailscale.attention',
-					source: 'tailscale-recovery',
-					message,
 				});
-			},
-			clearTailscaleAttention: () => {
-				safeTraceEvent(traceHandle, {
-					type: 'manual-diagnostic.tailscale.attention-cleared',
-					source: 'tailscale-recovery',
-				});
-			},
-			logTmuxAttachFailure: (tmuxResult) => {
-				safeTraceEvent(traceHandle, {
-					type: 'manual-diagnostic.tmux-attach-failed',
-					source: 'manual-diagnostic',
-					connection: {
-						connectionId: tmuxResult.connectionId,
-						tmuxSessionName: tmuxResult.tmuxSessionName,
-					},
-					details: {
-						tmuxAttachFailureReason: tmuxResult.tmuxAttachFailureReason,
-					},
-				});
-			},
-			logWarning: (message, error) => {
-				safeTraceEvent(traceHandle, {
-					type: 'manual-diagnostic.warning',
-					source: 'manual-diagnostic',
-					message,
-					error: serializeConnectionDiagnosticError(error),
-				});
-			},
-			trace: traceHandle,
-		});
 
-		return finish(traceHandle, result.connected ? 'connected' : 'failed', args);
+				return finish(
+					traceHandle,
+					result.connected ? 'connected' : 'failed',
+					args,
+				);
+			})(),
+			args.timeoutMs ?? DEFAULT_MANUAL_DIAGNOSTIC_TIMEOUT_MS,
+		);
 	} catch (error) {
 		if (!handle) {
 			throw error;
+		}
+		const isTimeout = error instanceof ManualDiagnosticTimeoutError;
+		if (isTimeout) {
+			safeTraceEvent(handle, {
+				type: 'manual-diagnostic.timeout',
+				source: 'manual-diagnostic',
+				message: error.message,
+				details: { timeoutMs: error.timeoutMs },
+			});
+			return finish(handle, 'failed', args);
 		}
 		safeTraceEvent(handle, {
 			type: 'manual-diagnostic.failed',
@@ -223,7 +292,10 @@ export async function runManualConnectionDiagnostic(
 		});
 		return finish(handle, 'failed', args);
 	} finally {
-		activeTraceHandle = null;
-		running = false;
+		if (activeRunToken === runToken) {
+			activeTraceHandle = null;
+			activeRunToken = null;
+			running = false;
+		}
 	}
 }
