+++ apps/mobile/src/lib/query-fns.ts
@@ -1,315 +1,19 @@
-// eslint-disable-next-line import/consistent-type-specifier-style -- Pure type import keeps Node integration tests from loading React Native.
-import type {
-	ConnectionDetails,
-	RnRussh,
-	SshConnection,
-	SshConnectionProgress,
-	SshShell,
-} from '@fressh/react-native-uniffi-russh';
-import {
-	serializeConnectionDiagnosticError,
-	type ConnectionDiagnosticConnectionIdentity,
-	type ConnectionDiagnosticEventInput,
-} from './connection-diagnostics';
+import { type SshConnectionProgress } from '@fressh/react-native-uniffi-russh';
+import { useMutation } from '@tanstack/react-query';
+import { useRouter } from 'expo-router';
+import { connectAndOpenShell } from './connect-and-open-shell';
 import { type InputConnectionDetails } from './connection-storage';
 import { rootLogger } from './logger';
-import { connectAndRememberConnection } from './ssh-connect-flow';
-import { extractTmuxAttachFailureReason } from './ssh-error-details';
-import { type RegisteredStartShellOptions } from './ssh-registry-store';
-import { AbortSignalTimeout } from './utils';
+import { useSshStore } from './ssh-store';
 
 const logger = rootLogger.extend('QueryFns');
-const DEFAULT_CONNECT_TIMEOUT_MS = 5_000;
-
-function diagnosticDisconnectTimeoutError(timeoutMs: number) {
-	return new Error(`Diagnostic SSH disconnect timed out after ${timeoutMs}ms`);
-}
-
-async function withDiagnosticDisconnectTimeout<T>(
-	promise: Promise<T>,
-	timeoutMs: number,
-): Promise<T> {
-	let timeoutId: ReturnType<typeof setTimeout> | null = null;
-	try {
-		return await Promise.race([
-			promise,
-			new Promise<never>((_, reject) => {
-				timeoutId = setTimeout(() => {
-					timeoutId = null;
-					reject(diagnosticDisconnectTimeoutError(timeoutMs));
-				}, timeoutMs);
-				const maybeNodeTimer = timeoutId as ReturnType<typeof setTimeout> & {
-					unref?: () => void;
-				};
-				maybeNodeTimer.unref?.();
-			}),
-		]);
-	} finally {
-		if (timeoutId !== null) clearTimeout(timeoutId);
-	}
-}
-
-export type ConnectAndOpenShellResult =
-	| {
-			status: 'connected';
-			sshConnection: SshConnection;
-			shellHandle: SshShell;
-			connectionId: string;
-			channelId: number;
-	  }
-	| {
-			status: 'tmux_attach_failed';
-			connectionId: string;
-			tmuxAttachFailureReason: string | null;
-			tmuxSessionName: string;
-			storedConnectionId: string;
-	  };
-
-type SaveConnection = (params: {
-	details: InputConnectionDetails;
-	priority: number;
-	label?: string;
-}) => Promise<unknown>;
-
-type ConnectTrace = {
-	event: (event: ConnectionDiagnosticEventInput) => void;
-};
-
-const getSecretsManager = async () => {
-	const { secretsManager } = await import('./secrets-manager');
-	return secretsManager;
-};
-
-const defaultSaveConnection: SaveConnection = async (params) => {
-	const secretsManager = await getSecretsManager();
-	return await secretsManager.connections.utils.upsertConnection(params);
-};
-
-// Shared resolver for turning stored details into a connect-ready security object.
-export async function resolveSecurityFromDetails(
-	connectionDetails: InputConnectionDetails,
-): Promise<ConnectionDetails['security']> {
-	const secretsManager = await getSecretsManager();
-	const privateKey = await secretsManager.keys.utils
-		.getPrivateKey(connectionDetails.security.keyId)
-		.then((e) => e.value);
-	return {
-		type: 'key',
-		privateKey,
-	};
-}
-
-// Shared connect flow used by both manual and silent auto-connect.
-export async function connectAndOpenShell(args: {
-	connectionDetails: InputConnectionDetails;
-	connect: typeof RnRussh.connect;
-	navigate: (params: { connectionId: string; channelId: number }) => void;
-	navigateWithError?: (params: {
-		connectionId: string;
-		tmuxAttachFailureReason: string | null;
-		tmuxSessionName: string;
-		storedConnectionId: string;
-	}) => void;
-	onConnectionProgress?: (progressEvent: SshConnectionProgress) => void;
-	abortSignalTimeoutMs?: number;
-	resolvedSecurity?: ConnectionDetails['security'];
-	saveConnection?: SaveConnection;
-	trace?: ConnectTrace;
-	diagnosticMode?: boolean;
-}): Promise<ConnectAndOpenShellResult> {
-	const {
-		connectionDetails,
-		connect,
-		navigate,
-		onConnectionProgress,
-		abortSignalTimeoutMs = DEFAULT_CONNECT_TIMEOUT_MS,
-		resolvedSecurity,
-		saveConnection = defaultSaveConnection,
-	} = args;
-	const traceEvent = (event: ConnectionDiagnosticEventInput) => {
-		try {
-			args.trace?.event(event);
-		} catch (error) {
-			logger.warn('Connect trace event failed', error);
-		}
-	};
-	const connectionIdentity: ConnectionDiagnosticConnectionIdentity = {
-		username: connectionDetails.username,
-		host: connectionDetails.host,
-		port: connectionDetails.port,
-		keyId: connectionDetails.security.keyId,
-		useTmux: connectionDetails.useTmux,
-		tmuxSessionName: connectionDetails.tmuxSessionName,
-	};
-	const security =
-		resolvedSecurity ?? (await resolveSecurityFromDetails(connectionDetails));
-
-	traceEvent({
-		type: 'ssh.connect.started',
-		source: 'saved-entry',
-		connection: connectionIdentity,
-	});
-	let rememberedConnection: {
-		sshConnection: SshConnection;
-		storedConnectionId: string;
-	};
-	try {
-		rememberedConnection = await connectAndRememberConnection({
-			connectionDetails,
-			connect,
-			saveConnection,
-			onConnectionProgress: (progressEvent) => {
-				logger.info('SSH connect progress event', progressEvent);
-				traceEvent({
-					type: 'ssh.connect.progress',
-					source: 'saved-entry',
-					connection: connectionIdentity,
-					details: { progressEvent },
-				});
-				onConnectionProgress?.(progressEvent);
-			},
-			abortSignalTimeoutMs,
-			resolvedSecurity: security,
-		});
-	} catch (error) {
-		traceEvent({
-			type: 'ssh.connect.failed',
-			source: 'saved-entry',
-			connection: connectionIdentity,
-			error: serializeConnectionDiagnosticError(error),
-		});
-		throw error;
-	}
-	const { sshConnection, storedConnectionId } = rememberedConnection;
-	const connectedIdentity = {
-		...connectionIdentity,
-		connectionId: sshConnection.connectionId,
-	};
-	traceEvent({
-		type: 'ssh.connect.connected',
-		source: 'saved-entry',
-		connection: connectedIdentity,
-		details: { storedConnectionId },
-	});
-	const cleanupDiagnosticConnection = async () => {
-		if (!args.diagnosticMode) return;
-		try {
-			await withDiagnosticDisconnectTimeout(
-				Promise.resolve(
-					sshConnection.disconnect?.({
-						signal: AbortSignalTimeout(abortSignalTimeoutMs),
-					}),
-				),
-				abortSignalTimeoutMs,
-			);
-			traceEvent({
-				type: 'ssh.diagnostic.disconnected',
-				source: 'saved-entry',
-				connection: connectedIdentity,
-			});
-		} catch (error) {
-			traceEvent({
-				type: 'ssh.diagnostic.disconnect-failed',
-				source: 'saved-entry',
-				connection: connectedIdentity,
-				error: serializeConnectionDiagnosticError(error),
-			});
-		}
-	};
-	let shellHandle: Awaited<ReturnType<typeof sshConnection.startShell>>;
-	try {
-		traceEvent({
-			type: 'ssh.shell.started',
-			source: 'saved-entry',
-			connection: connectedIdentity,
-		});
-		const startShellOptions: RegisteredStartShellOptions = {
-			term: 'Xterm',
-			useTmux: connectionDetails.useTmux,
-			tmuxSessionName: connectionDetails.tmuxSessionName,
-			abortSignal: AbortSignalTimeout(abortSignalTimeoutMs),
-			...(args.diagnosticMode ? { registerInStore: false } : {}),
-		};
-		shellHandle = await sshConnection.startShell(startShellOptions);
-	} catch (error) {
-		const tmuxAttachFailureReason = extractTmuxAttachFailureReason(error);
-		traceEvent({
-			type:
-				tmuxAttachFailureReason !== null
-					? 'ssh.shell.tmux-attach-failed'
-					: 'ssh.shell.failed',
-			source: 'saved-entry',
-			connection: connectedIdentity,
-			error: serializeConnectionDiagnosticError(error),
-			details: { tmuxAttachFailureReason, storedConnectionId },
-		});
-		if (tmuxAttachFailureReason !== null) {
-			if (!args.diagnosticMode) {
-				args.navigateWithError?.({
-					connectionId: sshConnection.connectionId,
-					tmuxAttachFailureReason,
-					tmuxSessionName: connectionDetails.tmuxSessionName,
-					storedConnectionId,
-				});
-			}
-			await cleanupDiagnosticConnection();
-			return {
-				status: 'tmux_attach_failed',
-				connectionId: sshConnection.connectionId,
-				tmuxAttachFailureReason,
-				tmuxSessionName: connectionDetails.tmuxSessionName,
-				storedConnectionId,
-			};
-		}
-		await cleanupDiagnosticConnection();
-		throw error;
-	}
-	traceEvent({
-		type: 'ssh.shell.connected',
-		source: 'saved-entry',
-		connection: connectedIdentity,
-		details: { channelId: shellHandle.channelId, storedConnectionId },
-	});
-
-	logger.info(
-		'Connected to SSH server',
-		sshConnection.connectionId,
-		shellHandle.channelId,
-	);
-	if (!args.diagnosticMode) {
-		navigate({
-			connectionId: sshConnection.connectionId,
-			channelId: shellHandle.channelId,
-		});
-	}
-
-	await cleanupDiagnosticConnection();
-
-	return {
-		status: 'connected',
-		sshConnection,
-		shellHandle,
-		connectionId: sshConnection.connectionId,
-		channelId: shellHandle.channelId,
-	};
-}
 
 export const useSshConnMutation = (opts?: {
 	onConnectionProgress?: (progressEvent: SshConnectionProgress) => void;
 }) => {
-	// Keep hook-only native dependencies lazy so connectAndOpenShell stays
-	// importable in Node integration tests.
-	/* eslint-disable @typescript-eslint/no-require-imports, @typescript-eslint/consistent-type-imports, react-compiler/react-compiler */
-	const { useMutation } =
-		require('@tanstack/react-query') as typeof import('@tanstack/react-query');
-	const { useRouter } = require('expo-router') as typeof import('expo-router');
-	const { useSshStore } =
-		require('./ssh-store') as typeof import('./ssh-store');
 	const router = useRouter();
 	const connect = useSshStore((s) => s.connect);
-	/* eslint-enable @typescript-eslint/no-require-imports, @typescript-eslint/consistent-type-imports, react-compiler/react-compiler */
 
-	// eslint-disable-next-line react-compiler/react-compiler -- useMutation is loaded lazily so connectAndOpenShell remains Node-importable.
 	return useMutation({
 		mutationFn: async (connectionDetails: InputConnectionDetails) => {
 			try {
